+++ src/PowerModelsITD.jl
@@ -69,6 +69,7 @@ module PowerModelsITD
     include("form/fotp.jl")
     include("form/lindist3flow.jl")
 
+    include("form_decomposition/boundary.jl")
     include("form_decomposition/acp.jl")
     include("form_decomposition/linear.jl")
     include("form_decomposition/lindist3flow.jl")